selftests: livepatch: Add new test to apply multiple livepatches
This test applies the same livepatch one after another as soon as the
transition finishes. It uses the same module file, but changes the
definitions inside the name to change it's name and how it is recognized
by the module loader.

As the test uses the same module file, it was necessary to adjust some
functions that check the current module/livepatch status. Before this
change, the module name was the same as the module filename, but now the
code get's the module name by checking the .modinfo section of the
module file.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>