livepatch: Move tests from lib/livepatch to selftests/livepatch
This code moving will allow writing more complex tests, like for example an
userspace C code that will call a livepatched kernel function. This change also
enables livepatch selftests to be exported using the "gen_tar", making it
possible to build and test the same source code on different kABI systems,
ensuring it's backward compatibility. The modules are now built as out-of-tree
modules, but being part of the kernel source means they will be maintained.

Another advantage of the code moving is to be able to easily change,
debug and rebuild the tests by running make on the selftests/livepatch directory,
which is not currently possible since the modules on lib/livepatch are
build and installed using the "modules" target.

The current approach also keeps the ability to execute the tests manually by
executing the scripts inside selftests/livepatch directory, as it's currently
supported. If the modules are modified, they needed to be rebuilt before running
the scripts though.

Having the modules being built as out-of-modules requires changing the currently
used 'modprobe' by 'insmod' and adapt the test scripts that check for the kernel
message buffer.

As there aren't any modules being built on lib/livepatch, remove the
TEST_LIVEPATCH Kconfig and it's references.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>