printk: Introduce register_console_force
The register_console_force function will register a console even if it wasn't
specified on boot. The new function will act like all consoles being registered
were using the CON_ENABLED flag, which is now obsolete.

The CON_ENABLED flag will be removed in the following patches and the drivers
that use it will migrate to register_console_force instead.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>