printk/nbcon: Use an enum to specify the required callback in console_is_usable()
The current usage of console_is_usable() is clumsy. The parameter @use_atomic
is boolean and thus not self-explanatory. The function is called twice in
situations when there are no-strict requirements.

Replace it with enum nbcon_write_cb which provides a more descriptive values
for all 3 situations.

Note that console_is_usable() checks only NBCON_USE_ATOMIC because .write_thread()
callback is mandatory. But the other two values still make sense because
they describe the intentions of the caller.

Suggested-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>