printk: Introduce console_is_nbcon
Besides checking if the current console is NBCON or not, console->flags
is also being read to in order to serve as argument of console_is_usable.

But the CON_NBCON flag is unique: it's set set just once in the console
registration and never cleared. In this case it can be possible to read
the flag when the console_srcu_lock is held (which is the case when
using for_each_console).

With this change it can be possible to remove the flags argument from
console_is_usable in the next patch, because there are cases where
the flags member is only used to check if the console is NBCON.

Signed-off-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>